// packages/siopv2/src/authorization-request/types.ts
export interface ClaimPayloadOptsVID1 extends ClaimPayloadCommonOpts {
  id_token?: IdTokenClaimPayload;
  vp_token?: PresentationDefinitionPayloadOpts;
}

export interface ClaimPayloadCommonOpts {
  // eslint-disable-next-line @typescript-eslint/no-explicit-any
  [x: string]: any;
}
// ...
interface AuthorizationRequestCommonOpts<CT extends ClaimPayloadCommonOpts> {
  // Yes, this includes common payload properties both at the payload level as well as in the requestObject.payload property. That is to support OAuth2 with or without a signed OpenID requestObject

  version: SupportedVersion;
  clientMetadata?: ClientMetadataOpts; // this maps to 'registration' for older SIOPv2 specs! OPTIONAL. This parameter is used by the RP to provide information about itself to a Self-Issued OP that would normally be provided to an OP during Dynamic RP Registration, as specified in {#rp-registration-parameter}.
  payload?: AuthorizationRequestPayloadOpts<CT>;
  requestObject: RequestObjectOpts<CT>;

  uriScheme?: Schema | string; // Use a custom scheme for the URI. By default openid:// will be used
}

export type AuthorizationRequestOptsVID1 = AuthorizationRequestCommonOpts<ClaimPayloadOptsVID1>;

export interface AuthorizationRequestOptsVD11 extends AuthorizationRequestCommonOpts<ClaimPayloadCommonOpts> {
  idTokenType?: string; // OPTIONAL. Space-separated string that specifies the types of ID token the RP wants to obtain, with the values appearing in order of preference. The allowed individual values are subject_signed and attester_signed (see Section 8.2). The default value is attester_signed.
}

export type CreateAuthorizationRequestOpts = AuthorizationRequestOptsVID1 | AuthorizationRequestOptsVD11;